+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **Unit tests fix**: FF now use auto-generated indexes too
 - **Unit tests fix**: offline detection so in this case, resetting local version number
 - avoid double synchronization with a protection, usefull in case of update > invalid_version wich may combine with a ping
+- use messages handlers to auto-synchronize after an option update
 - now use `api` instead of `ajax` in URL
 
 ### 0.7 ###